refactor: create shared 'useTablePagination' hook
