+++ src/components/Table.tsx
@@ -4,13 +4,10 @@ import {
   flexRender,
   getCoreRowModel,
   getSortedRowModel,
-  type PaginationState,
   type Row,
   type RowData,
-  type SortingState,
   useReactTable,
 } from "@tanstack/react-table";
-import { useMemo, useState } from "react";
 import { Link } from "react-router-dom";
 
 import { fetchPlayerList } from "../api/player/get/queries";
@@ -18,6 +15,7 @@ import { type IPlayer } from "../api/player/get/types";
 import { AQUA, NAVY } from "../constants/colors";
 import { tenMinutes } from "../constants/time";
 import { useDetectDarkTheme } from "../hooks/useDetectColorMode";
+import { useTablePagination } from "../hooks/useTablePagination";
 import { useAuthStore, useUserStore } from "../store";
 import { hyphenate } from "../utils/string";
 import { LoadingSpinner } from "./LoadingSpinner";
@@ -51,19 +49,14 @@ export const userColumnDefs = [
 ];
 
 export const Table = (): JSX.Element => {
-  const [sorting, setSorting] = useState<SortingState>([]);
-  const [{ pageIndex, pageSize }, setPagination] = useState<PaginationState>({
-    pageIndex: 0,
-    pageSize: 10,
-  });
-
-  const pagination = useMemo(
-    () => ({
-      pageIndex,
-      pageSize,
-    }),
-    [pageIndex, pageSize]
-  );
+  const {
+    sorting,
+    setSorting,
+    pagination,
+    setPagination,
+    pageIndex,
+    pageSize,
+  } = useTablePagination();
 
   const { getAuth } = useAuthStore();
   const auth = getAuth();
